Added population; CELE; overheat by duration.
By guessing at the fraction of spares assumed available, we find global overheat fractions that exceed this availability for each storm. We import land area and population to estimate the populations in these regions. There's now a plot which assesses the overheat fraction at each duration for a given geoelectric field too. Plots are saved and generally improved.